chore: resolve various clippy lints throughout the project
