+++ crates/ursa-index-provider/src/engine.rs
@@ -18,8 +18,7 @@ use anyhow::{anyhow, Error, Result};
 use axum::{
     body::Body,
     extract::Path,
-    http::StatusCode,
-    response::{IntoResponse, Response},
+    response::Response,
     routing::get,
     Extension, Json, Router,
 };
@@ -29,34 +28,17 @@ use fvm_ipld_blockstore::Blockstore;
 use libp2p::{gossipsub::TopicHash, identity::Keypair, multiaddr::Protocol, Multiaddr, PeerId};
 use std::{
     collections::VecDeque,
-    net::{IpAddr, Ipv4Addr, SocketAddr},
     str::FromStr,
     sync::Arc,
 };
 use tracing::{error, info, warn};
 use ursa_store::{BlockstoreExt, Dag, UrsaStore};
 use ursa_utils::convert_cid;
+use crate::provider::ProviderError;
 
 type CommandOneShotSender<T> = oneshot::Sender<Result<T, Error>>;
 type CommandOneShotReceiver<T> = oneshot::Receiver<Result<T, Error>>;
 
-pub enum ProviderError {
-    NotFoundError(Error),
-    InternalError(Error),
-}
-
-impl IntoResponse for ProviderError {
-    fn into_response(self) -> Response {
-        match self {
-            ProviderError::NotFoundError(e) => {
-                (StatusCode::NOT_FOUND, e.to_string()).into_response()
-            }
-            ProviderError::InternalError(e) => {
-                (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response()
-            }
-        }
-    }
-}
 // handlers
 async fn head<S: Blockstore + Store + Sync + Send + 'static>(
     Extension(state): Extension<Provider<S>>,
