+++ crates/ursa-index-provider/src/lib.rs
@@ -3,4 +3,6 @@ pub mod config;
 pub mod engine;
 pub mod provider;
 pub mod signed_head;
+
+#[cfg(test)]
 mod tests;
