+++ crates/ursa-network/src/codec/protocol.rs
@@ -135,28 +135,24 @@ impl RequestResponseCodec for UrsaExchangeCodec {
 mod tests {
     #[ignore = "todo"]
     #[tokio::test]
-    #[ignore = "todo"]
     async fn test_read_request() {
         todo!()
     }
 
     #[ignore = "todo"]
     #[tokio::test]
-    #[ignore = "todo"]
     async fn test_read_response() {
         todo!()
     }
 
     #[ignore = "todo"]
     #[tokio::test]
-    #[ignore = "todo"]
     async fn test_write_request() {
         todo!()
     }
 
     #[ignore = "todo"]
     #[tokio::test]
-    #[ignore = "todo"]
     async fn test_write_response() {
         todo!()
     }
