+++ crates/ursa-rpc-server/src/lib.rs
@@ -4,6 +4,8 @@ pub mod http;
 pub mod rpc;
 pub mod server;
 mod service;
-mod tests;
 
 pub use self::rpc::*;
+
+#[cfg(test)]
+mod tests;
