+++ crates/ursa-rpc-server/src/tests/api_test.rs
@@ -2,7 +2,6 @@
 mod tests {
     use crate::{
         api::{NetworkInterface, NodeNetworkInterface},
-        service,
         tests::{init, setup_logger},
     };
     use std::sync::Arc;
