+++ crates/ursa-store/src/store.rs
@@ -33,7 +33,7 @@ where
 /// Extension methods for inserting and retrieving IPLD data with CIDs
 pub trait BlockstoreExt: Blockstore {
     /// Get typed object from block store by CID
-    fn get_obj<T>(&self, cid: &Cid) -> anyhow::Result<Option<T>>
+    fn get_obj<T>(&self, cid: &Cid) -> Result<Option<T>>
     where
         T: DeserializeOwned,
     {
@@ -44,7 +44,7 @@ pub trait BlockstoreExt: Blockstore {
     }
 
     /// Put an object in the block store and return the Cid identifier.
-    fn put_obj<S>(&self, obj: &S, code: Code) -> anyhow::Result<Cid>
+    fn put_obj<S>(&self, obj: &S, code: Code) -> Result<Cid>
     where
         S: Serialize,
     {
@@ -53,14 +53,14 @@ pub trait BlockstoreExt: Blockstore {
     }
 
     /// Put raw bytes in the block store and return the Cid identifier.
-    fn put_raw(&self, bytes: Vec<u8>, code: Code) -> anyhow::Result<Cid> {
+    fn put_raw(&self, bytes: Vec<u8>, code: Code) -> Result<Cid> {
         let cid = Cid::new_v1(DAG_CBOR, code.digest(&bytes));
         self.put_keyed(&cid, &bytes)?;
         Ok(cid)
     }
 
     /// Batch put CBOR objects into block store and returns vector of CIDs
-    fn bulk_put<'a, S, V>(&self, values: V, code: Code) -> anyhow::Result<Vec<Cid>>
+    fn bulk_put<'a, S, V>(&self, values: V, code: Code) -> Result<Vec<Cid>>
     where
         Self: Sized,
         S: Serialize + 'a,
@@ -73,7 +73,7 @@ pub trait BlockstoreExt: Blockstore {
                 let cid = Cid::new_v1(DAG_CBOR, code.digest(&bytes));
                 Ok((cid, bytes))
             })
-            .collect::<anyhow::Result<Vec<_>>>()?;
+            .collect::<Result<Vec<_>>>()?;
 
         let cids = keyed_objects
             .iter()
